split components into separate folders
