+++ frontend/src/components/RoundView/HoleView.tsx
@@ -1,0 +1,3 @@
+export function HoleView({ id, par, strokes }: { id: number; par: number; strokes: number; }) {
+  return <div>Hole {id}: {strokes} / {par}</div>;
+}
