+++ frontend/src/components/RoundView/index.tsx
@@ -1,0 +1,5 @@
+import { RoundView } from "./RoundView";
+
+export {
+    RoundView,
+}
