+++ frontend/src/components/index.tsx
@@ -1,0 +1,7 @@
+import { CoursePicker } from "./CoursePicker";
+import { RoundView } from "./RoundView";
+
+export {
+    CoursePicker,
+    RoundView,
+}
